% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_request.R
\name{check_request}
\alias{check_request}
\title{Check if API request returns an error or success.}
\usage{
check_request(request)
}
\arguments{
\item{request}{Object returned by GET()}
}
\description{
Stop and print an error message if API call doesn't return 200 status code.
}
\keyword{internal}
