% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functor.R
\name{extract_functor}
\alias{extract_functor}
\title{Make a function that can be used to extract values from XML object.}
\usage{
extract_functor(x)
}
\arguments{
\item{x}{Either a document, a node set or a single node.}
}
\value{
\code{function(css)}
\itemize{
     \item css     :  Nodes to select.
}
}
\seealso{
Used by \code{\link{get_shopinfo_en}} and \code{\link{get_shopinfo_ja}}.
}
\keyword{internal}
