% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shopinfo_en.R
\name{get_shopinfo_en}
\alias{get_shopinfo_en}
\title{Scrape restaurant information (in English) from a given Tabelog URL.}
\usage{
get_shopinfo_en(shopURL)
}
\arguments{
\item{shopURL}{string, URL of restaurant, for example,
"https://tabelog.com/en/kyoto/A2601/A260301/26002222/".}
}
\value{
a data frame of restaurant info in English.
}
