% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shopinfo_ja.R
\name{get_shopinfo_ja}
\alias{get_shopinfo_ja}
\title{Scrape restaurant information (in Japanese) from a given Tabelog URL.

importFrom dplyr "%>%"}
\usage{
get_shopinfo_ja(shopURL)
}
\arguments{
\item{shopURL}{string, URL of restaurant, for example, 
"https://tabelog.com/kyoto/A2601/A260301/26002222/".}
}
\value{
a data frame of restaurant info in Japanese.
}
