% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shopURLs_ja.R
\name{get_shopURLs_ja}
\alias{get_shopURLs_ja}
\title{Extracts URLs of restaurants (in Japanese) listed on tabelog for a given city.}
\usage{
get_shopURLs_ja(city = "osaka", pages = 1:60)
}
\arguments{
\item{city}{string, lower cased city name in Japan.}

\item{pages}{integer vector, for example, 1:60.}
}
\value{
a character vector of URLs.
}
\description{
These URLs are sorted by the ratings of the restaurants in descending order.
}
